+++ core/src/main/java/io/cloudevents/core/impl/CloudEventReaderAdapter.java
@@ -24,7 +24,7 @@ public class CloudEventReaderAdapter implements CloudEventReader, CloudEventCont
 
     private final CloudEvent event;
 
-    CloudEventReaderAdapter(CloudEvent event) {
+    public CloudEventReaderAdapter(CloudEvent event) {
         this.event = event;
     }
