+++ core/src/main/java/io/cloudevents/core/v03/CloudEventBuilder.java
@@ -17,8 +17,8 @@
 package io.cloudevents.core.v03;
 
 import io.cloudevents.SpecVersion;
+import io.cloudevents.core.CloudEventUtils;
 import io.cloudevents.core.impl.BaseCloudEventBuilder;
-import io.cloudevents.core.impl.CloudEventUtils;
 import io.cloudevents.rw.CloudEventRWException;
 import io.cloudevents.types.Time;
